refactored the activities
